requirements.txt
# Update requirements.txt to include:
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
stripe>=8.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
drf-spectacular==0.27.1
django-cors-headers==4.3.1